refactor code organization, more comments, separate extra module

+++ lib/src/main/java/com/eyeem/recyclerviewtools/scroll_controller/_ScrollWithController.java
@@ -24,6 +24,6 @@ class _ScrollWithController extends AbstractController {
       } else {
          newY = config.view.getTranslationY() - dy;
       }
-      return config.limit((int) newY);
+      return config.limit(newY);
    }
 }
